+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## 1.0 branch
 ### 1.0.1-beta2
+* added `include` option for `FilesBackup` instances and `--include` option for
+    `FilesBackupCommand`;
+* no longer ignores hidden files by default;
 * added `FilesBackupCommandSubscriber` class, with command methods for events;
 * fixed little bug for `FilesBackup::getAllFiles()` method.